requirements.txt
asgiref==3.4.1
autopep8==1.5.7
beautifulsoup4==4.9.3
crispy-bootstrap5==0.4
Django==3.2.5
django-bootstrap-v5==1.0.4
django-crispy-forms==1.12.0
django-filter==2.4.0
django-tables2==2.4.0
djangorestframework==3.12.4
importlib-metadata==3.10.1
pycodestyle==2.7.0
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
zipp==3.5.0
